refactor zonecard
